+++ src/mysql_shell/clients/instance.py
@@ -2,7 +2,7 @@
 # See LICENSE file for licensing details.
 
 import logging
-from typing import Any
+from typing import Any, Collection, Mapping
 
 from ..builders import StringQueryQuoter
 from ..executors import BaseExecutor
@@ -14,7 +14,9 @@ from ..models.status import InstanceStatus
 
 logger = logging.getLogger()
 
-Attributes = dict[str, str] | None
+_Attributes = Mapping[str, str] | None
+_OptionalRoles = Collection[InstanceRole] | None
+_OptionalStates = Collection[InstanceStatus] | None
 
 
 class MySQLInstanceClient:
@@ -212,6 +214,18 @@ class MySQLInstanceClient:
         else:
             return [row["instance_name"] for row in rows]
 
+    def get_cluster_labels(self) -> list[str]:
+        """Gets the cluster labels."""
+        query = "SELECT cluster_name FROM mysql_innodb_cluster_metadata.clusters"
+
+        try:
+            rows = self._executor.execute_sql(query)
+        except ExecutionError:
+            logger.error("Failed to get cluster labels")
+            raise
+        else:
+            return [row["cluster_name"] for row in rows]
+
     def get_instance_replication_state(self) -> InstanceStatus:
         """Gets the instance replication state."""
         query = (
@@ -330,7 +344,7 @@ class MySQLInstanceClient:
             raise
 
     def search_instance_connections(self, name_pattern: str) -> list[int]:
-        """Searches the instance connections by name pattern."""
+        """Searches the instance connection IDs by name pattern."""
         query = (
             "SELECT processlist_id "
             "FROM performance_schema.threads "
@@ -408,7 +422,7 @@ class MySQLInstanceClient:
         else:
             return [Role.from_row(row["user"], row["host"]) for row in rows]
 
-    def search_instance_users(self, name_pattern: str, attrs: Attributes = None) -> list[User]:
+    def search_instance_users(self, name_pattern: str, attrs: _Attributes = None) -> list[User]:
         """Searches the instance users by name pattern and attributes."""
         attr_filter = "attribute LIKE {string}"
         attr_substr = '%"{key}": "{val}"%'
@@ -461,7 +475,7 @@ class MySQLInstanceClient:
     def stop_instance_processes(self, process_ids: list[int]) -> None:
         """Kills the instances processes by ID."""
         query = "KILL CONNECTION {id}"
-        queries = [query.format(id=self._quoter.quote_value(str(pid))) for pid in process_ids]
+        queries = [query.format(id=self._quoter.quote_value(pid)) for pid in process_ids]
         queries = ";".join(queries)
 
         try: